+++ src/ipred.rs
@@ -4,6 +4,6 @@ pub unsafe extern "C" fn get_upsample(
     angle: libc::c_int,
     is_sm: libc::c_int,
 ) -> libc::c_int {
-    return (angle < 40 as libc::c_int && wh <= 16 as libc::c_int >> is_sm)
+    return (angle < 40 && wh <= 16 >> is_sm)
         as libc::c_int;
 }
